requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
sympy==1.12
randfacts==0.20.2
aiohttp==3.8.6
dadjokes==1.3.2